+++ lib/src/view/view_model.dart
@@ -1,0 +1,11 @@
+import 'package:queryflow/queryflow.dart';
+
+class ViewModel {
+  final String name;
+  final SelectBuilderFetch query;
+
+  ViewModel({
+    required this.name,
+    required this.query,
+  });
+}