+++ lib/src/view/view_syncronizer.dart
@@ -1,0 +1,77 @@
+import 'package:queryflow/src/executor/executor.dart';
+import 'package:queryflow/src/logger/query_logger.dart';
+import 'package:queryflow/src/view/view_model.dart';
+
+class ViewSyncronizer {
+  final List<ViewModel> views;
+  final String databaseName;
+  final QueryLogger? logger;
+  final Executor executor;
+  ViewSyncronizer({
+    required this.views,
+    required this.databaseName,
+    this.logger,
+    required this.executor,
+  });
+
+  Future<void> syncronize() async {
+    logger?.i('Start syncronizing view');
+    for (var view in views) {
+      final viewExists = await _viewExists(view.name);
+      if (!viewExists) {
+        await _createView(view);
+        logger?.s("Created view '${view.name}'");
+      } else {
+        final existingColumns = await _getViewColumns(view.name);
+        final currentQuery = view.query.toSql();
+        bool hasChanges = false;
+        for (var column in existingColumns) {
+          if (!currentQuery.contains(column)) {
+            hasChanges = true;
+            break;
+          }
+        }
+        if (hasChanges) {
+          await _updateView(view);
+          logger?.s("Updated view '${view.name}'");
+        }
+      }
+    }
+    logger?.i('Finished syncronizing views');
+  }
+
+  Future<bool> _viewExists(String name) async {
+    final result = await executor.execute(
+      '''SELECT TABLE_NAME 
+        FROM INFORMATION_SCHEMA.VIEWS 
+        WHERE TABLE_SCHEMA = '$databaseName' 
+        AND TABLE_NAME = '$name';''',
+    );
+    return result.isNotEmpty;
+  }
+
+  Future<void> _createView(ViewModel view) async {
+    final query = view.query.toSql();
+    await executor.executePrepared(
+      '''CREATE OR REPLACE VIEW ${view.name} AS ?;''',
+      [query],
+    );
+  }
+
+  Future<void> _updateView(ViewModel view) async {
+    final query = view.query.toSql();
+    executor.executePrepared(
+      '''CREATE OR REPLACE VIEW ${view.name} AS ?;''',
+      [query],
+    );
+  }
+
+  Future<List<String>> _getViewColumns(String name) {
+    return executor.executePrepared('''SELECT COLUMN_NAME 
+        FROM INFORMATION_SCHEMA.COLUMNS 
+        WHERE TABLE_SCHEMA = '$databaseName' 
+        AND TABLE_NAME = ?;''', [name]).then((result) {
+      return result.map((e) => e['COLUMN_NAME'].toString()).toList();
+    });
+  }
+}